refactor: flatten utils directory
